requirements.txt
openai==1.12.0
anthropic==0.18.0
streamlit==1.29.0
pyyaml==6.0
pydantic==2.5.0
click==8.1.7
requests==2.31.0
pandas==2.1.0